Check in pristine copy of registry client package
We're forking from Heroku's package to make some breaking API changes.
Checking in a clean copy first so that we have an easy-to-see history
for future commits.

Added some copyright headers to existing files, since the licenses
differ and so we can more easily identify which files are covered under
which license.

[Updates ch3708]